Updated EpiProject
I managed to get the graph to pull up but it does not seem to be plotting anything or it is plotting but only in a straight line. All I did was on Line 39 I removed the variable "results=" and moved lines 52 - 61 into the sir_model() def. I am still having trouble getting it to plot right any ideas why the graph looks weird from the data?